Breakup of index and db files into smaller pieces
